Make sure JpaAfterClearFlusher is registered if an EntityManager exists
Issue #6 when JPA is in play, make sure to register the JpaAfterClearFlusher with a valid EntityManager. This will be added to the flush-chain, to be potentially executed when flushAfterClear is turned on. This can be either set on BeanCollection or the override configuration.

Note that @PersistenceContext cannot be used in the auto-configuration, since this will fail the tests. Instead optional @Autowired is used.