Removed direct dependency on Spring Security
The AuthenticationManager declared in BeanMapperAutoConfig.java caused
applications without Spring Security on their classpath to fail when
including the beanmapper-spring-boot-starter as dependency. Since the
usage of the Spring Security features of BeanMapper-spring is optional,
projects without Spring Security should still be supported in
beanmapper-spring-boot-starter. In such cases, the @BeanLogicSecured and
@BeanRoleSecured annotations are not processed

Fixes #10